+++ common/GM_NET/PDA/Inc/net_pda_pid_init.h
@@ -17,8 +17,8 @@ uint8_t *net_pda_init_sub_pid_status_ctl(uint8_t step, uint8_t status);
 uint8_t *net_pda_init_sub_pid_status_rsp(uint8_t step, uint8_t status);
 uint8_t *net_pda_init_sub_pid_status_rqt(uint8_t step);
 
-uint8_t *net_pda_init_sub_pid_absolute_battery_ctl(uint32_t save_data);
-uint8_t *net_pda_init_sub_pid_absolute_battery_rsp(uint32_t save_data);
+uint8_t *net_pda_init_sub_pid_absolute_battery_ctl(int64_t save_data);
+uint8_t *net_pda_init_sub_pid_absolute_battery_rsp(int64_t save_data);
 uint8_t *net_pda_init_sub_pid_absolute_battery_rqt(void);
 
 uint8_t *net_pda_init_sub_pid_move_sensor_ctl(uint8_t mode);
